Don't flag the wheel as a "manylinux" wheel
Manylinux is a pep8 policy that assert that the wheels are usable on a
variety of Linux machines that satisfy standard requirements.

In order for a wheel to be manylinux compatible, it must respect some
requirements as well. They are described in
https://www.python.org/dev/peps/pep-0513/#the-manylinux1-policy

The current snap wheels do not satisfy these requirements (see #122), and
therefore cannot be flagged as Manylinux compatible.

In the wait of a solution that patches the wheel, this PR removes the incorrect flagging.